% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsi_cube-methods.R
\name{dim.hsi_cube}
\alias{dim.hsi_cube}
\title{Get Dimensions of an hsi_cube}
\usage{
\method{dim}{hsi_cube}(x)
}
\arguments{
\item{x}{An \link{hsi_cube} object.}
}
\value{
Integer vector \code{c(rows, cols, bands)}.
}
\description{
Get Dimensions of an hsi_cube
}
\examples{
cube <- hs_example_cube()
dim(cube)

}
