% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{hs_dark_correct}
\alias{hs_dark_correct}
\title{Apply Dark Current Correction}
\usage{
hs_dark_correct(cube, dark)
}
\arguments{
\item{cube}{An \link{hsi_cube} object.}

\item{dark}{An \link{hsi_cube} object or 3D array representing the dark reference.}
}
\value{
An \link{hsi_cube} object with dark-corrected values.
}
\description{
Subtracts a dark reference from the cube data.
}
\examples{
cube <- hs_simulate_cube(rows = 10, cols = 10, noise_sd = 0)
dark <- hsi_cube(array(0.01, dim(cube$data)), cube$wavelengths)
corrected <- hs_dark_correct(cube, dark)

}
