% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{hs_mnf}
\alias{hs_mnf}
\title{Minimum Noise Fraction Transform}
\usage{
hs_mnf(cube, n_components = 5L)
}
\arguments{
\item{cube}{An \link{hsi_cube} object.}

\item{n_components}{Integer. Number of components. Default \code{5}.}
}
\value{
A list with class \code{"hsi_mnf"} (same structure as \code{\link[=hs_pca]{hs_pca()}}).
}
\description{
Orders components by signal-to-noise ratio rather than variance.
Estimates noise covariance from spatial first-differences.
}
\examples{
cube <- hs_example_cube()
mnf <- hs_mnf(cube, n_components = 3)
dim(mnf$scores)

}
