% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{hs_msc}
\alias{hs_msc}
\title{Multiplicative Scatter Correction}
\usage{
hs_msc(cube, reference = NULL)
}
\arguments{
\item{cube}{An \link{hsi_cube} object.}

\item{reference}{Numeric vector. Reference spectrum. Default \code{NULL} = mean
spectrum of the cube.}
}
\value{
An \link{hsi_cube} object with MSC-corrected spectra.
}
\description{
Corrects spectra for multiplicative and additive scatter effects by
regressing each spectrum against a reference spectrum (default: mean
spectrum).
}
\examples{
cube <- hs_example_cube()
msc_cube <- hs_msc(cube)

}
