% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{hs_ndi}
\alias{hs_ndi}
\title{Compute Normalized Difference Index (General Purpose)}
\usage{
hs_ndi(cube, band1, band2)
}
\arguments{
\item{cube}{An \link{hsi_cube} object.}

\item{band1}{Numeric. Center wavelength or range \code{c(min, max)} for first band.}

\item{band2}{Numeric. Center wavelength or range \code{c(min, max)} for second band.}
}
\value{
A numeric matrix, in \verb{[-1, 1]} for nonnegative input spectra.
Zero denominators are NA; transformed signed spectra can exceed these bounds.
}
\description{
NDI = (R_band1 - R_band2) / (R_band1 + R_band2). A flexible building block
for any two-band ratio index.
}
\examples{
cube <- hs_example_cube()
ndi <- hs_ndi(cube, band1 = 540, band2 = 660)
range(ndi, na.rm = TRUE)

}
