% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{hs_pca}
\alias{hs_pca}
\title{PCA on Hyperspectral Cube}
\usage{
hs_pca(cube, n_components = 5L, center = TRUE, scale = FALSE)
}
\arguments{
\item{cube}{An \link{hsi_cube} object.}

\item{n_components}{Integer. Number of components to retain. Default \code{5}.}

\item{center}{Logical. Center bands before PCA. Default \code{TRUE}.}

\item{scale}{Logical. Scale bands to unit variance. Default \code{FALSE}.}
}
\value{
A list with class \code{"hsi_pca"}:
\describe{
\item{scores}{3D array (rows x cols x n_components) of component score maps.}
\item{loadings}{Matrix (bands x n_components) of spectral loadings.}
\item{variance_explained}{Numeric vector of proportion of variance per component.}
\item{center}{Centering vector used (or \code{FALSE}).}
\item{scale}{Scaling vector used (or \code{FALSE}).}
\item{wavelengths}{Wavelength vector from input cube.}
}
}
\description{
Performs Principal Component Analysis on the spectral dimension of an HSI
cube. Reduces the spectral bands to a smaller number of orthogonal
components ordered by explained variance.
}
\examples{
cube <- hs_example_cube()
pca <- hs_pca(cube, n_components = 3)
pca$variance_explained

}
