% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hs_plot_rgb}
\alias{hs_plot_rgb}
\title{Synthesize RGB Image from Spectral Cube}
\usage{
hs_plot_rgb(cube, r = 640, g = 550, b = 460, stretch = "linear")
}
\arguments{
\item{cube}{An \link{hsi_cube} object.}

\item{r, g, b}{Numeric. Center wavelengths for R, G, B channels.
Defaults: r=640, g=550, b=460.}

\item{stretch}{Character. Histogram stretch: \code{"linear"} (default), \code{"none"}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Creates a pseudo-color RGB composite by mapping three wavelength bands
to the red, green, and blue channels.
}
\examples{
cube <- hs_example_cube()
hs_plot_rgb(cube)

}
