% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{hs_process}
\alias{hs_process}
\title{Execute a Processing Recipe}
\usage{
hs_process(cube, recipe = hs_recipe(), learn = TRUE)
}
\arguments{
\item{cube}{An \link{hsi_cube} object.}

\item{recipe}{A recipe from \code{\link[=hs_recipe]{hs_recipe()}}.}

\item{learn}{Logical. Allow fitting an unspecified MSC reference on this cube.
Set FALSE for validation or prediction data.}
}
\value{
List containing the processed \code{cube} and reusable fitted \code{recipe}.
}
\description{
Execute a Processing Recipe
}
