% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-read.R
\name{hs_read_cube}
\alias{hs_read_cube}
\title{Read a Hyperspectral Data Cube from Any Supported Format}
\usage{
hs_read_cube(path, ...)
}
\arguments{
\item{path}{Path to the hyperspectral data file.}

\item{...}{Additional arguments passed to the format-specific reader.}
}
\value{
An \link{hsi_cube} object.
}
\description{
Auto-detects format from file extension and dispatches to the appropriate
reader. Supported formats: ENVI (.hdr), TIFF (.tif/.tiff), Cubert (.cu3s).
}
\examples{
hdr_path <- hs_example_files()
cube <- hs_read_cube(hdr_path, verbose = FALSE)
dim(cube)

}
