% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-read.R
\name{hs_read_tiff}
\alias{hs_read_tiff}
\title{Read a Multi-Channel TIFF File}
\usage{
hs_read_tiff(path, wavelengths, fwhm = NULL, verbose = TRUE)
}
\arguments{
\item{path}{Path to multi-channel TIFF file.}

\item{wavelengths}{Numeric vector of wavelengths (required for TIFF files
as they lack spectral metadata).}

\item{fwhm}{Numeric vector of FWHM values. Default \code{NULL}.}

\item{verbose}{Logical. Print progress messages. Default \code{TRUE}.}
}
\value{
An \link{hsi_cube} object.
}
\description{
Reads a multi-band TIFF file as a hyperspectral cube. Requires the \code{terra}
package for TIFF reading.
}
\examples{
\donttest{
# Requires terra package
# cube <- hs_read_tiff("path/to/image.tif", wavelengths = seq(430, 910, by = 8))
}

}
