% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{hs_resample}
\alias{hs_resample}
\title{Resample Spectra to New Wavelength Grid}
\usage{
hs_resample(cube, target_wavelengths, method = "linear")
}
\arguments{
\item{cube}{An \link{hsi_cube} object.}

\item{target_wavelengths}{Numeric vector. Target wavelength positions in nm.}

\item{method}{Character. Interpolation method: \code{"linear"} (default),
\code{"spline"}.}
}
\value{
An \link{hsi_cube} object with resampled spectra.
}
\description{
Interpolates the spectral data to a new set of wavelength positions.
}
\examples{
cube <- hs_example_cube()
resampled <- hs_resample(cube, target_wavelengths = seq(450, 900, by = 10))
length(resampled$wavelengths)

}
