% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{hs_roi_polygon}
\alias{hs_roi_polygon}
\title{Define Polygon ROI}
\usage{
hs_roi_polygon(cube, vertices)
}
\arguments{
\item{cube}{An \link{hsi_cube} object.}

\item{vertices}{Matrix with columns \code{x} and \code{y} (polygon vertices in pixel
coordinates).}
}
\value{
A list with class \code{"hsi_roi"} containing the mask.
}
\description{
Creates a polygon region of interest using point-in-polygon testing.
}
\examples{
cube <- hs_example_cube()
verts <- matrix(c(5, 5, 15, 15, 5, 20, 5, 20), ncol = 2,
                dimnames = list(NULL, c("x", "y")))
roi <- hs_roi_polygon(cube, verts)

}
