% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{hs_snv}
\alias{hs_snv}
\title{Standard Normal Variate Correction}
\usage{
hs_snv(cube)
}
\arguments{
\item{cube}{An \link{hsi_cube} object.}
}
\value{
An \link{hsi_cube} object with SNV-corrected spectra.
}
\description{
Normalizes each spectrum to zero mean and unit variance. Useful for
reducing multiplicative scatter effects and baseline variation.
}
\examples{
cube <- hs_example_cube()
snv_cube <- hs_snv(cube)

}
