% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsi_cube-methods.R
\name{print.hsi_cube}
\alias{print.hsi_cube}
\title{Print an hsi_cube Object}
\usage{
\method{print}{hsi_cube}(x, ...)
}
\arguments{
\item{x}{An \link{hsi_cube} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisible \code{x}.
}
\description{
Displays a compact summary of the hyperspectral cube.
}
\examples{
cube <- hs_example_cube()
print(cube)

}
