% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{scale_color_wavelength}
\alias{scale_color_wavelength}
\alias{scale_colour_wavelength}
\title{Wavelength-to-Color Scale for Spectral Plots}
\usage{
scale_color_wavelength(...)

scale_colour_wavelength(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradientn()}}.}
}
\value{
A \link[ggplot2:Scale]{ggplot2::Scale} object.
}
\description{
Maps wavelength values to approximate visible light colors for spectral
line coloring. Returns a ggplot2 continuous color scale.
}
\examples{
cube <- hs_example_cube()
library(ggplot2)
df <- data.frame(
  wavelength = cube$wavelengths,
  value = cube$data[15, 15, ]
)
ggplot(df, aes(wavelength, value, color = wavelength)) +
  geom_line() +
  scale_color_wavelength()

}
