% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsi_cube-methods.R
\name{[.hsi_cube}
\alias{[.hsi_cube}
\title{Subset an hsi_cube Object}
\usage{
\method{[}{hsi_cube}(x, i, j, k, ...)
}
\arguments{
\item{x}{An \link{hsi_cube} object.}

\item{i}{Row indices (spatial).}

\item{j}{Column indices (spatial).}

\item{k}{Band indices (spectral).}

\item{...}{Additional arguments (ignored).}
}
\value{
A subsetted \link{hsi_cube} object.
}
\description{
Extract spatial and/or spectral subsets from a cube. Subsetting preserves
the \code{hsi_cube} class.
}
\examples{
cube <- hs_example_cube()
sub <- cube[1:10, 1:10, 1:5]
dim(sub)

}
