% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsi_cube-methods.R
\name{as_tibble.hsi_cube}
\alias{as_tibble.hsi_cube}
\title{Convert hsi_cube to Tibble}
\usage{
as_tibble.hsi_cube(x, ..., long = TRUE, .name_repair = "unique")
}
\arguments{
\item{x}{An \link{hsi_cube} object.}

\item{...}{Additional arguments (ignored).}

\item{long}{Logical. If \code{TRUE} (default), returns long format. See
\code{\link[=as.data.frame.hsi_cube]{as.data.frame.hsi_cube()}} for details.}

\item{.name_repair}{Name repair strategy (passed to \code{\link[tibble:as_tibble]{tibble::as_tibble()}}).}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Convert hsi_cube to Tibble
}
\examples{
cube <- hs_example_cube()
tb <- as_tibble.hsi_cube(cube[1:3, 1:3, 1:3])
head(tb)

}
